Fixes #1698 

This PR encompasses a few small optimisations:

- Use `.specific(defer=True)` when getting pages. This means there's no need to fetch the specific page models at all, removing the extra queries which comes with it.
- Use the DB to calculate the latest `lastmod`, rather than looping through all pages.
- Use a slightly simpler / faster way of determining the last modified date when looping through pages (copied from Django)
- Cache the Wagtail site on the view

The query drops in the tests is deceptive. These changes dropped them by 2 in the tests - the rest are from using a non-database cache.

With ~10000 pages (averaged over 10 requests using `gunicorn` and `oha`):

Before:

```
Slowest:      1.4849 secs
Fastest:      1.0741 secs
Average:      1.1292 secs
Requests/sec: 0.8856
```

35 DB queries

After:

```
Slowest:      1.8591 secs
Fastest:      1.4128 secs
Average:      1.4964 secs
Requests/sec: 0.6683
```

18 DB queries

--- 

I'm not sure why the updated version appears slower - although I think it's within the margin of error. It's running fewer queries, and `defer=True` should make the codepaths much simpler, too. The tests were run on `bakerydemo` and I suspect with more than just a single model the improvements may be greater still.