@RealOrangeOne 2024-06-25
The cause of the slowdown seems to be related to the use of `.specific(defer=True)`. As this is the main optimisation, I'll do some digging as to why it's so much slower.