@RealOrangeOne 2024-06-25
The issue is indeed from `.specific(defer=True)`, particularly when used as a queryset (as opposed to a singular use of `.specific_deferred`). I have a rough fix, but it'll pollute this PR too much, so I'll open it as a separate one. This shouldn't be merged until that's in.